requirements.txt
fastapi
psycopg2-binary
uvicorn[standard]
pika
requests
psycopg2-binary